+++ volmdlr/nurbs/operations.py
@@ -642,8 +642,8 @@ def split_surface_v(obj, param, **kwargs):
     temp_obj = insert_knot_func(obj, [None, param], num=[0, insertion_count], check_num=False)
 
     # Knot vectors
-    knotvectors = helper_split_knot_vectors(temp_obj.degree_v, temp_obj.knots_vector_v, temp_obj.nb_v, param,
-                                            span_func)
+    knotvectors = helper_split_knot_vectors(temp_obj.degree_v, temp_obj.knots_vector_v,
+                                            temp_obj.nb_v, param, span_func)
     return construct_split_surfaces(temp_obj, knotvectors, "v", knot_span, insertion_count)
 
 
@@ -825,7 +825,7 @@ def decompose_surface(obj, return_params, **kwargs):
 
     :param obj: surface
     :type obj: BSplineSurface3D
-        :param return_params: If True, returns the parameters from start and end of each Bézier patch with repect to the
+    :param return_params: If True, returns the parameters from start and end of each Bézier patch with repect to the
      input curve.
     :type return_params: bool
     :return: a list of Bezier patches