+++ volmdlr/shells.py
@@ -2,6 +2,7 @@
 import math
 import random
 import warnings
+import traceback
 from itertools import chain, product
 from typing import Any, Dict, Iterable, List, Tuple, Union
 
@@ -801,6 +802,7 @@ class Shell3D(volmdlr.core.CompositePrimitive3D):
         """
         meshes = []
         for i, face in enumerate(self.faces):
+            # print(face.name)
             try:
                 face_mesh = face.triangulation()
 
@@ -808,6 +810,7 @@ class Shell3D(volmdlr.core.CompositePrimitive3D):
                 face_mesh = None
                 warnings.warn(f"Could not triangulate {face.__class__.__name__} with index {i} in the shell "
                               f"{self.name} faces. Probably because topology error in contour2d.")
+                print(traceback.format_exc())
                 continue
             if face_mesh:
                 meshes.append(face_mesh)